+++ src/App.tsx
@@ -23,7 +23,7 @@ import { getUsersById } from "./api/index";
 import AuthGuard from "./guard/AuthGuard";
 import AdressConfirmationModal from "./components/blocks/AdressConfirmationModal/AdressConfirmationModal";
 import Orders from "./pages/Orders";
-import SummeryOrder from "./components/elements/SummeryOrder";
+// import SummeryOrder from "./components/elements/SummeryOrder";
 
 const App = withTranslation()(function App() {
   const dispatch = useDispatch();
@@ -93,7 +93,7 @@ const App = withTranslation()(function App() {
                 path={routes.favoriteRestaurants}
                 element={<FavoriteRestaurants />}
               />
-               <Route path={`${routes.summeryorder}/:id`} element={<SummeryOrder/>}/>
+               {/* <Route path={`${routes.summeryorder}/:id`} element={<SummeryOrder/>}/> */}
                <Route path={routes.orderList}element={<Orders/>} />
             </Route>
             <Route path={routes.home} element={<Home />} />
